// angular/src/app/prospect/models/prospect.model.ts
export interface Prospect {
  idProspect: number;
  idUser: number;
  nomeResponsavel: any;
	idGenero: number;
	data: string;
	nomeContato: string;
	emailContato: string;
	telFixo: string;
	telCel: string;
  cpfCliente: string;
	nomeCliente: string;
	dataNascimento: string;
	peso: string;
	altura: string;
	proposta: string;
	historico: string;
	dtInsert: string;
}